Simplify proofs of work
Proofs of work now use 32-byte challenges and the work to be proved is:
generate a BLAKE2b hash of Challenge||N that starts with K zeros for
some N.  N is then the proof of work, while K is the difficulty.
Obviously this is very time- and memory-easy for the server to verify.