Fix and replace key exchange mechanisms
Monocypher has deprecated the crypto_key_exchange_*() functions.  In
place of these functions, it recommends using a "higher level key
exchange protocol".  In most of the places that these are used, we
already use a higher level key exchange protocol, and in fact should
have been using simple crypto_x25519 in the first place, as the secret
values produced by crypto_key_exchange were being fed into a
BLAKE2b-based HDKF anyway.

In one place - header encryption of initial peer-to-peer messages - we
used the result of crypto_key_exchange() directly as a shared encryption
key.  This isn't terrible, but according to Monocypher it isn't ideal.
In place of crypto_key_exchange we now use a simple key exchange
protocol where we use BLAKE2b(raw_shared_secret || ika || ikb), as
recommended by Monocypher.  This probably isn't ideal, but it doesn't
need to be.  Encryption of these headers is just for disguising the
protocol being used: everything that is encrypted with this key is
public information, and the key is thrown away immediately.  If one of
the long-term identity keys is leaked, then an attacker can indeed
decrypt these headers.  But all they will find out is that it is an
initial message, the initial (public) ephemeral key used by Alice, and the
(public) signed prekey and (public) one-time prekey chosen by Alice for
the initial key exchange with Bob.  As I understand it, in many
algorithms these choices are sent over the public network unencrypted
anyway.  We hide them on a 'better safe than sorry' basis and to avoid
the server being able to tell when sessions are being initiated.