Create unified server context struct
struct server_ctx holds the persistent state of the server.  Separating
this into its own object will make it more clear what is ephemeral and
what is persistent when it comes to persisting state to disk between
sessions in the future.  It also reduces the number of parameters that
need to be passed to packet handling functions.