Remove pointless reply-to-msn concept
The idea was that each reply/ack would include the MSN (message sequence
number) of the message it was a reply to/acknowledgement of.  However,
msns are reused far too often for this to be a useful concept.
Fundamentally, they exist at the wrong level.  The higher level
messaging protocol needs its own concept of message identifiers to tie
requests/messages to replies/acks.