Move MSG_FORWARD handling into its own function
This is the last bit of message handling code in the server to be moved
into its own function.

This commit also changes the errg() macro to take the label as a
parameter.  This makes it more obvious that it involves jumping to a
label, and allows its use where naming the label `fail' would be
misleading.