Disable {de/}registering of stacks on yield
It appears that valgrind does not like what I am doing here.  I suspect
that the way I have implemented it is incorrect.  I think that what I
should probably be doing is using VALGRIND_STACK_CHANGE to set the size
of an inactive stack to 0, then to set it to the correct size again
later.

At least for the time being, stacks are deregistered only when the fibre
finishes.  This makes some sense: the memory is valid, just not in
active use as a stack, until the stack returns, and then it is and
should be invalid.