packet: Allow the specification of cv in prepare()
This is preparation for supporting RESUME messages.  A RESUME message
will require that the client and server gracefully shut down their last
session using a GOODBYE exchange.  Whoever shuts down the connection
with a GOODBYE message (usually the client) will include a random
32-byte value in the message.  The GOODBYE_ACK message (usually sent by
the server) will similarly include a random 32-byte value.  Later, when
the client wishes to reconnect to the server, the client will include in
its HELLO message the random value specified by the server and the
server will include the client's random value in its REPLY.

The benefit of this design is that it ensures that resumption is
indistinguishable from a completely fresh HELLO/REPLY session to an
external attacker, even one that knows that HELLO and REPLY messages are
being exchanged.