Switch to a fibre-per-peer model in the server
The server now spins off a low-priority fibre for processing handshakes.
When that fibre successfully completes a handshake with a peer, it spins
off a fibre for handling incoming packets from that peer.  The
datagram-handling fibre only calls recvfrom and puts the received
datagram into the correct queue (hshakeq or peer->recvq).